Add page view count functionality to website
A new JavaScript file has been created to fetch the current page view count from an external API, parse the JSON response, and display the count on the webpage. An element has been added to index.html to display this count and the script has been linked to this file.